[FIX] 코드 리뷰 반영 및 gateway 모듈 도입 (#304) (#346)
* fix: import 구조 수정 및 Facade 역할 분리 - #304

- 성능 저하를 우려한 .* 와 같은 형식의 import 제거
- Facade를 각 역할 별로 분리

* fix: 이미지 S3 업로드 성능 개선 - #304

- 리사이징 알고리즘 수정
  - SCALE_SMOOTH: Java에서 가장 느린 방식
  - Graphics2D & RenderingHints 기반 Bilinear 보간법을 통한 CPU 연산 속도 향상
- 이미지 별 (변환 -> 리사이징 -> 업로드) 과정을 병령 파이프라인 적용

* refactor: 상품 동기화 트랜잭션 적용 수정 및 외부 API 실패에 따른 정합성 확보 - #304

- eventLogUseCase에서만 트랜잭션을 적용하여 불필요한 DB 커넥션 풀 점유 제거
- API 호출 롤백 불가에 따른 트랜잭션 외부에서 진행
- Retry & DLT를 통한 API 호출 실패에 따른 이벤트 정합성 확보

* fix: 상품 옵션 값의 유니크 제약 수정 - #304

- 옵션 그룹과 값을 통합하여 유니크 제약 설정

* fix: 브랜드 및 카테고리 중복 검증 로직 수정 - #304

- 불필요한 대량 조회 제거

* fix: 잘못된 쿼리 문법 수정 - #304

* fix: 브랜드 및 카테고리 조회 시, 페이징 적용 - #304

* refactor: 스웨거 API 명세서 수정 - #304

* fix: 기본 KafkaTemplate 설정 도입 - #304

* feat: gateway 모듈 빌드 추가 - #304

* fix: 페이지네이션 응답에서의 잘못된 데이터 검증 수정 - #304

* fix: 브랜드 이름 중복 검사 시, 소문자로 검증 - #304

- 불필요한 전환을 제거하고 소문자로 변환하여 DB의 데이터의 LOWERCASE와 검증

* fix: 브랜드 및 카테고리 이름 중복 로직 오류 수정 - #304

- 자기 자신을 제외한 중복 여부 검증할 수 있도록 수정

* fix: 페이지네이션 타입에 맞는 요청 및 응답 타입 수정 - #304

+++ product/src/test/java/com/back/product/adapter/in/ApiV1ProductQueryControllerTest.java
@@ -190,16 +190,16 @@ class ApiV1ProductQueryControllerTest {
     class FindSimilarProductsTest {
 
         private final Long PRODUCT_INFO_ID = 1L;
-        private final Long DEFAULT_COUNT = 5L;
-        private final Long CUSTOM_COUNT = 8L;
-        private final Long INVALID_COUNT = 51L;
+        private final Integer DEFAULT_COUNT = 5;
+        private final Integer CUSTOM_COUNT = 8;
+        private final Integer INVALID_COUNT = 51;
 
         private ProductSearchResponseDto createMockProductSearchResponseDto(Long count) {
             return ProductSearchResponseDto.builder()
                     .products(Collections.singletonList(ProductSearchDto.builder().productName("Similar Product").build()))
                     .totalElements(count)
-                    .totalPages(1L)
-                    .currentPage(0L)
+                    .totalPages(1)
+                    .currentPage(0)
                     .build();
         }
 
@@ -207,12 +207,12 @@ class ApiV1ProductQueryControllerTest {
         @DisplayName("성공: size 파라미터 없이 유사 상품을 조회한다 (기본값 사용)")
         void findSimilarProducts_success_defaultCount() throws Exception {
             PageRequestDto request = PageRequestDto.builder()
-                    .page(0L)
+                    .page(0)
                     .size(DEFAULT_COUNT)
                     .build();
 
             // given
-            ProductSearchResponseDto mockResponseDto = createMockProductSearchResponseDto(DEFAULT_COUNT);
+            ProductSearchResponseDto mockResponseDto = createMockProductSearchResponseDto(DEFAULT_COUNT.longValue());
             given(productFacade.findSimilarProducts(PRODUCT_INFO_ID, request))
                     .willReturn(mockResponseDto);
 
@@ -230,12 +230,12 @@ class ApiV1ProductQueryControllerTest {
         @DisplayName("성공: size 파라미터를 지정하여 유사 상품을 조회한다")
         void findSimilarProducts_success_customCount() throws Exception {
             PageRequestDto request = PageRequestDto.builder()
-                    .page(0L)
+                    .page(0)
                     .size(CUSTOM_COUNT)
                     .build();
 
             // given
-            ProductSearchResponseDto mockResponseDto = createMockProductSearchResponseDto(CUSTOM_COUNT);
+            ProductSearchResponseDto mockResponseDto = createMockProductSearchResponseDto(CUSTOM_COUNT.longValue());
             given(productFacade.findSimilarProducts(PRODUCT_INFO_ID, request))
                     .willReturn(mockResponseDto);
 
@@ -255,7 +255,7 @@ class ApiV1ProductQueryControllerTest {
         void findSimilarProducts_failure_invalidCount() throws Exception {
             // given - no need to mock facade as validation happens before facade call
             PageRequestDto request = PageRequestDto.builder()
-                    .page(0L)
+                    .page(0)
                     .size(INVALID_COUNT)
                     .build();
